rename from HURUMap to HURUmap
